# swagger.yaml
openapi: 3.0.0
info:
  title: Bootcam API
  description: A full-featured REST API for bootcamp directory management
  version: 1.0.0
  contact:
    name: Mubarak
    url: https://github.com/muubaarik/Bootcam-Api
servers:
  - url: https://bootcam-api-4hmt.onrender.com/api/v1
    description: Production
  - url: http://localhost:5001/api/v1
    description: Local Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /bootcamps:
    get:
      tags: [Bootcamps]
      summary: Get all bootcamps
      description: |
        Supports filtering, sorting, pagination.

        **Filter examples:**
        - `?housing=true`
        - `?averageCost[lte]=10000`
        - `?careers[in]=Web Development`

        **Sort:** `?sort=-createdAt`

        **Select fields:** `?select=name,description`

        **Pagination:** `?page=1&limit=5`
      parameters:
        - in: query
          name: select
          schema:
            type: string
          example: name,description
        - in: query
          name: sort
          schema:
            type: string
          example: -createdAt
        - in: query
          name: page
          schema:
            type: integer
          example: 1
        - in: query
          name: limit
          schema:
            type: integer
          example: 5
        - in: query
          name: housing
          schema:
            type: boolean
      responses:
        "200":
          description: List of bootcamps

    post:
      tags: [Bootcamps]
      summary: Create a bootcamp
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, description, address, careers]
              properties:
                name:
                  type: string
                  example: My Coding Bootcamp
                description:
                  type: string
                  example: Learn to code fast
                website:
                  type: string
                  example: https://mybootcamp.com
                phone:
                  type: string
                  example: "(617) 555-1234"
                email:
                  type: string
                  example: info@mybootcamp.com
                address:
                  type: string
                  example: 233 Bay State Rd Boston MA 02215
                careers:
                  type: array
                  items:
                    type: string
                  example: ["Web Development"]
                housing:
                  type: boolean
                  example: true
                jobAssistance:
                  type: boolean
                  example: true
                jobGuarantee:
                  type: boolean
                  example: false
                acceptGi:
                  type: boolean
                  example: true
      responses:
        "201":
          description: Bootcamp created
        "401":
          description: Not authorized
        "400":
          description: Validation error
...
  /users:
    get:
      tags: [Users]
      summary: Get all users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
        "403":
          description: Forbidden

    post:
      tags: [Users]
      summary: Create user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [user, publisher, admin]
      responses:
        "201":
          description: User created
...